block: bfq: fix Oops at system shutdown
Commit fbca70d33072 ("Port commit ae11889636: turn blkg_lookup_create into
blkg_lookup") introduced the possibility of bfq_find_set_group returning a
NULL pointer. When this happened in __bfq_bic_change_cgroup,
it caused an Oops a few lines later. This Oops was reported as reliably
happening on one machine on the sync_bfqq path at system shutdown.

This patch restores the prior behavior of falling back to using the
bfqd->root_group if bfqg is NULL. The behavior is restored, but at
a different place.

Note Well: This patch assumes that the prior behavior was correct.
     	   That assumption may not be warranted.

Signed-off-by: Lee Tibbert <LeeTibbert@users.noreply.github.com>
Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>